requirements.txt
langchain==1.1.3
langchain-core==1.1.3
langchain-community==0.4.1

# langchain
# langchain-core
# langchain-community
langchain-pinecone==0.2.13
flask==3.1.1
sentence-transformers==4.1.0
pypdf==6.4.1
python-dotenv==1.0.0
langchain-groq
langchain_huggingface
streamlit
-e .